core: split libvlc loading code paths per platform
# Conflicts:
#	src/LibVLCSharp/Core.cs
#	src/LibVLCSharp/Core/Constants.cs
#	src/LibVLCSharp/Core/Core.Android.cs
#	src/LibVLCSharp/Core/Core.Apple.cs
#	src/LibVLCSharp/Core/Core.Desktop.cs
#	src/LibVLCSharp/Core/Core.UWP.cs
#	src/LibVLCSharp/Core/Core.cs
